refactor: features 기반 → app 기반 colocation 구조로 마이그레이션
- src/features/home/* → app/(home)/로 이동
- src/atoms/* → app/(home)/atoms/로 통합
- src/features/posts/PostRenderer.tsx → app/posts/[slug]/로 이동
- src/components/common/Giscus.tsx → app/posts/[slug]/로 이동
- 빈 src/features/, src/components/common/ 폴더 삭제
- import 경로를 상대 경로로 변경

Closes #56
