+++ app/posts/[slug]/PostRenderer.tsx
