--- src/atoms/categories/index.ts
@@ -1,7 +1,0 @@
-import { atom } from 'jotai';
-
-import { Category } from '@/interfaces';
-
-export const categoriesAtom = atom<Category[]>([]);
-
-export const selectedCategoryAtom = atom<string>('');
