--- src/atoms/profile/index.ts
@@ -1,3 +1,0 @@
-import { atom } from 'jotai';
-
-export const profileImageAtom = atom<string>('');
